Refs #37785 - Race condition during container push
