+++ app/lib/actions/katello/repository/create_root.rb
@@ -2,23 +2,13 @@ module Actions
   module Katello
     module Repository
       class CreateRoot < Actions::EntryAction
-        def plan(root, relative_path = nil)
-          begin
-            root.save!
-          rescue ActiveRecord::RecordInvalid => e
-            if root.is_container_push && e.message.include?("Container Repository Name") && e.message.include?("conflicts with an existing repository")
-              logger.warn("Skipping repository creation as container push repository already exists: #{root.container_push_name}")
-              return
-            end
-            raise e
-          end
+        def plan(root)
+          root.save!
           repository = ::Katello::Repository.new(:environment => root.organization.library,
-                                      :content_view_version => root.organization.library.default_content_view_version,
-                                      :root => root)
-          repository.container_repository_name = relative_path if root.docker? && root.is_container_push
-          repository.relative_path = relative_path || repository.custom_repo_path
+                                                 :content_view_version => root.organization.library.default_content_view_version,
+                                                 :root => root)
+          repository.relative_path = repository.custom_repo_path
           repository.save!
-
           action_subject(repository)
           plan_action(::Actions::Katello::Repository::Create, repository)
         end
